# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for demonstration purposes.",
        "title": "My API",
        "contact": {},
        "version": "1.0"
    },
    "host": "8.218.102.112:4880",
    "basePath": "/api/v1",
    "paths": {
...
        "/transaction/sol/get_swap_request_status": {
            "get": {
                "description": "Get swap request status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Get swap request status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "swap_transaction",
                        "name": "swap_transaction",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SwapTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/transaction/sol/send_swap_transaction": {
            "get": {
                "description": "Send swap request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Send swap request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "transaction",
                        "name": "transaction",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SwapTransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/{chainType}/{tradeType}/get_swap_route": {
            "get": {
                "description": "Get swap route for a specific token address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "swap"
                ],
                "summary": "Get swap route",
                "parameters": [
                    {
                        "enum": [
                            "sol",
                            "eth",
                            "btc"
                        ],
                        "type": "string",
                        "description": "Chain type",
                        "name": "chainType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "tx"
                        ],
                        "type": "string",
                        "description": "Trade type",
                        "name": "tradeType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "sol",
                            "eth",
                            "btc"
                        ],
                        "type": "string",
                        "description": "Token in chain",
                        "name": "token_in_chain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "sol",
                            "eth",
                            "btc"
                        ],
                        "type": "string",
                        "description": "Token out chain",
                        "name": "token_out_chain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "From address",
                        "name": "from_address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Slippage",
                        "name": "slippage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token in address",
                        "name": "token_in_address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token out address",
                        "name": "token_out_address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "In amount",
                        "name": "in_amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Fee",
                        "name": "fee",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Is anti MEV",
                        "name": "is_anti_mev",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Legacy",
                        "name": "legacy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Token total supply",
                        "name": "token_total_supply",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Initial real token reserves",
                        "name": "initial_real_token_reserves",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Initial virtual sol reserves",
                        "name": "initial_virtual_sol_reserves",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Initial virtual token reserves",
                        "name": "initial_virtual_token_reserves",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SwapRouteData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.Instruction": {
            "type": "object",
            "properties": {
                "accounts": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "data": {
                    "type": "string"
                },
                "programIdIndex": {
                    "type": "integer"
                }
            }
        },
...
        "response.Quote": {
            "type": "object",
            "properties": {
                "inAmount": {
                    "type": "string"
                },
                "inDecimals": {
                    "type": "integer"
                },
                "inputMint": {
                    "type": "string"
                },
                "otherAmountThreshold": {
                    "type": "string"
                },
                "outAmount": {
                    "type": "string"
                },
                "outDecimals": {
                    "type": "integer"
                },
                "outputMint": {
                    "type": "string"
                },
                "platformFee": {
                    "type": "integer"
                },
                "priceImpactPct": {
                    "type": "string"
                },
                "routePlan": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.RoutePlan"
                    }
                },
                "slippageBps": {
                    "type": "integer"
                },
                "swapMode": {
                    "type": "string"
                },
                "timeTaken": {
                    "type": "number"
                }
            }
        },
        "response.RawTx": {
            "type": "object",
            "properties": {
                "lastValidBlockHeight": {
                    "type": "integer"
                },
                "prioritizationFeeLamports": {
                    "type": "integer"
                },
                "recentBlockhash": {
                    "type": "string"
                },
                "swapTransaction": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "response.RoutePlan": {
            "type": "object",
            "properties": {
                "percent": {
                    "type": "integer"
                },
                "swapInfo": {
                    "$ref": "#/definitions/response.SwapInfo"
                }
            }
        },
        "response.Signature": {
            "type": "object",
            "properties": {
                "signature": {
                    "description": "Use a map to capture numeric keys",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "response.SwapInfo": {
            "type": "object",
            "properties": {
                "ammKey": {
                    "type": "string"
                },
                "feeAmount": {
                    "type": "integer"
                },
                "feeMint": {
                    "type": "string"
                },
                "inAmount": {
                    "type": "string"
                },
                "inputMint": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "outAmount": {
                    "type": "string"
                },
                "outputMint": {
                    "type": "string"
                }
            }
        },
        "response.SwapRouteData": {
            "type": "object",
            "properties": {
                "amountInUSD": {
                    "type": "string"
                },
                "amountOutUSD": {
                    "type": "string"
                },
                "jito_order_id": {},
                "quote": {
                    "$ref": "#/definitions/response.Quote"
                },
                "raw_tx": {
                    "$ref": "#/definitions/response.RawTx"
                }
            }
        },
        "response.SwapTransactionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "base64SwapTransaction": {
                            "type": "string"
                        },
                        "lastValidBlockHeight": {
                            "type": "integer"
                        },
                        "swapTransaction": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "$ref": "#/definitions/response.TransactionMessage"
                                },
                                "signatures": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/response.Signature"
                                    }
                                }
                            }
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.TransactionMessage": {
            "type": "object",
            "properties": {
                "accountKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "header": {
                    "type": "object",
                    "properties": {
                        "numReadonlySignedAccounts": {
                            "type": "integer"
                        },
                        "numReadonlyUnsignedAccounts": {
                            "type": "integer"
                        },
                        "numRequiredSignatures": {
                            "type": "integer"
                        }
                    }
                },
                "indexToProgramIds": {
                    "type": "object",
                    "additionalProperties": true
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Instruction"
                    }
                },
                "recentBlockhash": {
                    "type": "string"
                }
            }
        },
...
    }
}